+++ internal/konnect/controlplanes/client.go
@@ -18,6 +18,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/kong/kubernetes-ingress-controller/v3/internal/manager/metadata"
 	"github.com/oapi-codegen/runtime"
 	openapi_types "github.com/oapi-codegen/runtime/types"
 )
@@ -379,6 +380,11 @@ func NewClient(server string, opts ...ClientOption) (*Client, error) {
 	if client.Client == nil {
 		client.Client = &http.Client{}
 	}
+	client.RequestEditors = append([]RequestEditorFn{func(ctx context.Context, req *http.Request) error {
+		req.Header.Set("User-Agent", metadata.UserAgent())
+		return nil
+	}}, client.RequestEditors...)
+
 	return &client, nil
 }